requirements.txt
tensorflow==2.10.0
numpy==1.23.5
pandas>=2.2.3
scikit-learn>=1.6.1
xgboost>=2.1.4
uvicorn>=0.27.0
fastapi>=0.109.0
pillow>=10.4.0
requests>=2.32.3
pydantic==1.10.15
python-multipart>=0.0.9